pyproject.toml
[project]
name = "bot-telegram"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "psutil>=7.2.2",
    "python-telegram-bot>=22.7",
]
